Move collection containers to their own directory.
