+++ scripts/containers/collection/CollectionEditPage.js
@@ -1,8 +1,8 @@
 import { bindActionCreators } from "redux";
 import { connect } from "react-redux";
 
-import CollectionEdit from "../components/collection/CollectionEdit";
-import * as BucketActions from "../actions/bucket";
+import CollectionEdit from "../../components/collection/CollectionEdit";
+import * as BucketActions from "../../actions/bucket";
 
 
 function mapStateToProps(state) {
