+++ scripts/containers/collection/CollectionHistoryPage.js
@@ -2,9 +2,9 @@ import { bindActionCreators } from "redux";
 import { connect } from "react-redux";
 import { push as updatePath } from "react-router-redux";
 
-import CollectionHistory from "../components/collection/CollectionHistory";
-import * as CollectionActions from "../actions/collection";
-import * as NotificationsActions from "../actions/notifications";
+import CollectionHistory from "../../components/collection/CollectionHistory";
+import * as CollectionActions from "../../actions/collection";
+import * as NotificationsActions from "../../actions/notifications";
 
 
 function mapStateToProps(state) {
