+++ scripts/routes.js
@@ -18,10 +18,10 @@ import BucketHistoryPage from "./containers/bucket/BucketHistoryPage";
 import GroupCreatePage from "./containers/GroupCreatePage";
 import GroupEditPage from "./containers/GroupEditPage";
 import GroupHistoryPage from "./containers/GroupHistoryPage";
-import CollectionListPage from "./containers/CollectionListPage";
-import CollectionHistoryPage from "./containers/CollectionHistoryPage";
-import CollectionCreatePage from "./containers/CollectionCreatePage";
-import CollectionEditPage from "./containers/CollectionEditPage";
+import CollectionListPage from "./containers/collection/CollectionListPage";
+import CollectionHistoryPage from "./containers/collection/CollectionHistoryPage";
+import CollectionCreatePage from "./containers/collection/CollectionCreatePage";
+import CollectionEditPage from "./containers/collection/CollectionEditPage";
 import AddFormPage from "./containers/AddFormPage";
 import BulkFormPage from "./containers/BulkFormPage";
 import EditFormPage from "./containers/EditFormPage";
